SLVS-1564 Move IActiveConfigScopeTracker  outside binding folder
