+++ src/IssueViz.UnitTests/OpenInIDE/OpenInIdeConfigScopeValidatorTests.cs
@@ -19,7 +19,7 @@
  */
 
 using SonarLint.VisualStudio.Core;
-using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.ConfigurationScope;
 using SonarLint.VisualStudio.IssueVisualization.OpenInIde;
 using SonarLint.VisualStudio.TestInfrastructure;
