+++ src/SLCore.IntegrationTests/RuleDescriptionConversionSmokeTest.cs
@@ -20,7 +20,7 @@
 
 using System.Windows.Documents;
 using SonarLint.VisualStudio.Core;
-using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.ConfigurationScope;
 using SonarLint.VisualStudio.Education.Layout.Logical;
 using SonarLint.VisualStudio.Education.Rule;
 using SonarLint.VisualStudio.Education.XamlGenerator;
