+++ src/SLCore.Listeners.UnitTests/Implementation/ConnectedModeSuggestionListenerTests.cs
@@ -19,7 +19,7 @@
  */
 
 using SonarLint.VisualStudio.ConnectedMode.Binding.Suggestion;
-using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.ConfigurationScope;
 using SonarLint.VisualStudio.SLCore.Core;
 using SonarLint.VisualStudio.SLCore.Listener.Binding;
