+++ src/SLCore.Listeners/Implementation/ListFilesListener.cs
@@ -21,7 +21,7 @@
 using System.ComponentModel.Composition;
 using System.IO;
 using SonarLint.VisualStudio.Core;
-using SonarLint.VisualStudio.Core.Binding;
+using SonarLint.VisualStudio.Core.ConfigurationScope;
 using SonarLint.VisualStudio.SLCore.Common.Helpers;
 using SonarLint.VisualStudio.SLCore.Core;
 using SonarLint.VisualStudio.SLCore.Listener.Files;
